# api/menufamille/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "MenuFamille",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/info",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/recette/CreateRecipeAllInfo": {
      "post": {
        "tags": [
          "recipe"
        ],
        "summary": "Create a recipe",
        "description": "Create a recipe and link it with it's denree, categories and tags",
        "operationId": "Create_Recipe_All_Infos",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "recipe",
            "in": "body",
            "description": "full information of a recipe",
            "required": true,
            "type": "object",
            "properties": {
              "nom" : {"type" : "string"},
              "difficulte" : {"type" : "integer", "format": "int64"},
              "calorie" : {"type" : "integer", "format": "int64"},
              "temps_cuisson" : {"type" : "integer", "format": "int64"},
              "temps_preparation" : {"type" : "integer", "format": "int64"},
              "nb_personne" : {"type" : "integer", "format": "int64"},
              "nutriscore" : {"type" : "string"},
              "preparation" : {"type" : "string"},
              "tags": { "$ref":"#/definitions/Tags_full"},
              "categorie": { "$ref": "#/definitions/Categorie_full"},
              "denree" : { "$ref": "#/definitions/Denree_full"}
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                      "id_recette": {"type": "integer","format": "int64"},
                      "nom" : {"type" : "string"},
                      "difficulte" : {"type" : "integer", "format": "int64"},
                      "calorie" : {"type" : "integer", "format": "int64"},
                      "temps_cuisson" : {"type" : "integer", "format": "int64"},
                      "temps_preparation" : {"type" : "integer", "format": "int64"},
                      "nb_personne" : {"type" : "integer", "format": "int64"},
                      "nutriscore" : {"type" : "string"},
                      "preparation" : {"type" : "string"}
                  }
                }
              }
            }
          }
        },
...
      },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header"
    }
  },
  "definitions": {
...
    "Tags_full":{
      "type": "object",
      "properties": {
        "id_tag": { "type" : "integer", "format": "int64"},
        "nom" : {"type" : "string"}
      }
    },
    "Categorie_full":{
      "type": "object",
      "properties": {
        "id_categorie": { "type" : "integer", "format": "int64"},
        "nom" : {"type" : "string"}
      }
    },
...
    "Denree_full" :{
      "type": "object",
      "properties": {
        "id_denree" : {"type" : "integer", "format": "int64"},
        "nom" : {"type" : "string"},
        "nutriscore" : {"type" : "string"},
        "calories": {"type": "integer","format": "int64"},
        "recette_denree" : {"type": "object", "properties": {"quantite" : {"type" : "integer", "format": "int64"}}}
      }
    },
...
  }
}